added turn functionality to drive; added touch sensor for manual ex. shutdown

+++ src/components/Drivable.java
@@ -1,6 +1,6 @@
 package components;
 
 public interface Drivable {
-    void drive(int speed, float turn);
+    void drive(int speed, int turn);
     void rotateOnPlace(int speed, int degree);
 }
